Refactor app structure for Postgres migration and observability
- Split app logic into app/main.go, app/db.go, app/handlers.go, app/models.go
- Keep main.go minimal with metrics middleware and route wiring
- Add Postgres-ready DB query functions with Prometheus timing
- Separate handlers and models for cleaner code organization
